🐞 fix: Add space to fix dependancy name parsing (#50)
## Description

We've been ping for review where we shouldn't have been. The bump here
is `@opentelemetry/exporter-trace-otlp-proto from 0.33.0 to 0.34.0` with
`@remix-run/*:minor` as `npm-disallowlist`.

After investigation I found the reason, the dependancy name ends with
`to`, therefore, when parsing the PR's title to get the `to` version, it
takes `to from 0.33.0` and extracts the digits of the `from` again.

The fix is simple: I added a space before the target: ` to` instead of
`to`.

## Related Issue

https://linear.app/fewlines/issue/TRB-171/fix-auto-merge

## Motivation and Context

We don't want 🐞🐛🐝 of any kind (at least here)

## How Has This Been Tested?

I added a test case.

## Types of changes

- ~Chore (non-breaking change which refactors / improves the existing
code base)~
- Bug fix (non-breaking change which fixes an issue)
- ~New feature (non-breaking change which adds functionality)~
- ~Breaking change (fix or feature that would cause existing
functionality to
  change)~

## Checklist:

- 🔴 My change requires a change to the documentation.
- 🔴 I have updated the documentation accordingly.
- ✅ I have added tests to cover my changes.
- ✅ All new and existing tests passed.
- ✅ I have read the
[CONTRIBUTING](https://github.com/fewlinesco/eslint-config/blob/main/CONTRIBUTING.md)
document.